Refactored code and moved filtering functionality to get_router.py.
Code that goes through handler's arguments and filters the routers is
now in get_router.py. As a result, the handlers will now focus solely on
creating their respective documents. Timestamps for relays and bridges
have been added, but more work needs to be done for the bandwidth
handler, as well as some of the fields in the details handler.