Added sqlite database functionality:
Created a database to store the information of each router according to
the different fields. One table (summary) contains all the fields except
for 'flags' and 'addresses'. A database and the summary table is created
using database(). get_summary_routers() takes in the arguments from
SummaryHandler, reads the summary table, and appends the necessary
routers to a tuple of lists that will be passed back to the SummaryHandler.
'flags' and 'addresses' will be in their own separate tables. The next step
will be to add these two tables to the database.