Externalized GET request parameter parsing.
Added handlers/arguments.py, which has a single function parse.  parse
takes the GET request parameter dictionary as an argument, and returns
a dictionary mapping strings to values.  That dictionary can be used
as the keyword arguments dictionary for the database module functions,
so now the database module does not need to understand anything about
GET requests.