SQLite database updated automatically according to a specified interval.
We use a Timer object that updates the database every
database.DB_UPDATE_INTERVAL seconds.  As a Timer, this is not performed
in the main thread, so doesn't block incoming connections.  Happily,
the database update is automatically a single transaction, so the
update is not visible to clients until it is complete.